requirements.txt
anyio==4.0.0
appdirs==1.4.4
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.0
contourpy==1.2.0
cycler==0.12.1
dnspython==2.4.2
fonttools==4.46.0
forex-python==1.8
frozendict==2.3.9
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
h11==0.14.0
html5lib==1.1
httpcore==0.18.0
httplib2==0.22.0
httpx==0.25.0
idna==3.4
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.2
multitasking==0.0.11
nanoid==2.0.0
numpy==1.26.0
packaging==23.2
pandas==2.1.1
peewee==3.17.0
Pillow==10.0.1
plotly==5.18.0
protobuf==4.24.4
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pylance==0.8.5
pymongo==4.6.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-telegram-bot==20.6
python-telegram-bot-pagination==0.0.3
pytz==2023.3.post1
qrcode==7.4.2
removebg==0.4
requests==2.31.0
rsa==4.9
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tenacity==8.2.3
tradermade==0.8.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
webencodings==0.5.1
websockets==8.1
yfinance==0.2.32
